requirements.txt
# Core Web Framework (Python 3.12+ recommended for Snowflake wheels)
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# HTTP & API (>=2.32.4 required by snowflake-connector-python)
requests>=2.32.4,<3

# Environment & Configuration
python-dotenv==1.0.0

# Background Jobs & Scheduling
apscheduler==3.10.4

# Utilities
werkzeug==3.0.0
packaging==23.2
typing-extensions==4.8.0

# Snowflake (backend venv: py -3.12 -m venv backend/venv312)
snowflake-connector-python>=3.10.1
snowflake-sqlalchemy>=1.5.1
sqlalchemy>=2.0.36